fix: type error in post return
